amended propertys models.py product model price to floatfield to fix bug with decimal place of json data numberic value

+++ propertys/models.py
@@ -21,7 +21,7 @@ class Product(models.Model):
     sku = models.CharField(max_length=254, null=True, blank=True)
     name = models.CharField(max_length=254)
     description = models.TextField()
-    price = models.DecimalField(max_digits=15, decimal_places=2)
+    price = models.FloatField()
     rating = models.DecimalField(max_digits=6, decimal_places=2, null=True, blank=True)
     image = models.ImageField(blank=True)
